% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\docType{data}
\name{age_labs}
\alias{age_labs}
\title{Will add documentation}
\format{
An object of class \code{character} of length 8.
}
\usage{
age_labs
}
\description{
Will add documentation
}
\keyword{datasets}
