% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_functions.R
\name{cm_filter}
\alias{cm_filter}
\title{#'  Scales contact matrices
#'
#' Take the ratio of the max eigenvalues of two matrices
#'
#' @param x square matrix with the same dimentions as y
#' @param y square matrix with the same dimentions as y
#'
#' @export
#'
Create a contact matrix with filters}
\usage{
cm_filter(
  survey,
  countries = "United Kingdom",
  age_limits = c(0, 5, 18, 30, 40, 50, 60, 70),
  symmetric = TRUE,
  filter_text = list(),
  boots = 1,
  return_matrix = TRUE
)
}
\arguments{
\item{survey}{a socialmixR survey object}

\item{countries}{a character vector passed to socialmixR::contact_matrix}

\item{filter_text}{= a list of text to filter the survey, passed to
socialmixR::contact_matrix}

\item{boots}{a numerical, defaulted to 1, indicating the number of bootsraps,
passed to socialmixR::contact_matrix
#' @param return_matrix a logical, defaulted to true, to return a socialmixR
contact matrix object. Returns a socialmixR survey object if set to FALSE}

\item{age.limit}{a numeric vectors used to create breaks of ages, passed to
socialmixR::contact_matrix}
}
\description{
#'  Scales contact matrices
#'
#' Take the ratio of the max eigenvalues of two matrices
#'
#' @param x square matrix with the same dimentions as y
#' @param y square matrix with the same dimentions as y
#'
#' @export
#'
Create a contact matrix with filters
}
