% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cm.R
\name{create_cm}
\alias{create_cm}
\title{Create a contacts matrix using socialmixr}
\usage{
create_cm(
  survey,
  countries = "United Kingdom",
  age_limits,
  symmetric = FALSE,
  filter_text = list(),
  boots = 1,
  return_matrix = TRUE,
  ...
)
}
\arguments{
\item{survey}{a survey object from socialmixr}

\item{age_limits}{lower limits of the age groups over which to construct the matrix}

\item{symmetric}{whether to make matrix symmetric}

\item{boots}{number of bootstrap sample}

\item{export_var_names}{Whether to export the variable names into a seperate file}

\item{matrix_output}{return just the matrix or the matrix and participants default = TRUE}

\item{country}{limit to one or more countries; if not given, will use all countries in the survey; these can be given as country names or 2-letter (ISO Alpha-2) country codes}

\item{filter}{list object for filtering the contacts matrix to work}
}
\value{
Either contact matrix or contact matrix and with participants information
}
\description{
Wrapper function for socialmix which helps quickly create filtered contacts matrices
and reshape and returns multiple objects with the prefix table
which relate to different questions asked in the survey
}
