% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_scaling_matrices_observed.R
\name{create_scaling_matrices_observed}
\alias{create_scaling_matrices_observed}
\title{Will add documentation}
\usage{
create_scaling_matrices_observed(
  comix_survey,
  polymod_survey,
  nboots,
  matrices_path,
  file_name,
  age_limits = c(0, 5, 18, 30, 40, 50, 60, 70),
  age_limits_sym = c(18, 30, 40, 50, 60, 70)
)
}
\description{
Will add documentation
}
