% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\name{gg_matrix}
\alias{gg_matrix}
\title{Plots a data.table of converted socialmixR objects with ggplot}
\usage{
gg_matrix(dt, breaks = c(0, 0.5, 1), age_lab = FALSE)
}
\arguments{
\item{dt}{data.table object converted from socialmixR contact matrix objects}

\item{breaks}{numeric vector passed to ggplot::scale_fill_gradientn breaks
parameter}

\item{age_lab}{character vector of age labels passed to
ggplot::scale_x_discrete labels parameter}
}
\description{
This function takes a data.table object converted from socialmixR contact
matrix objects from the socialmixR package by study
}
