% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_functions.R
\name{impute_cm}
\alias{impute_cm}
\title{Impute Comix matrix based on Polymod}
\usage{
impute_cm(
  comix,
  polymod,
  i = 0,
  impute_rows = 1:3,
  impute_cols = 1:3,
  observed_rows = 4:9,
  observed_cols = 1:9,
  reciprocol = TRUE,
  ...
)
}
\arguments{
\item{comix}{comix matrix which might need to be split}

\item{polymod}{polymod matrix}

\item{i}{default as 0 for when single matrices}

\item{impute_rows}{rows to impute}

\item{impute_cols}{cols to impute}

\item{observed_rows}{rows that are observed}

\item{observed_cols}{cols that are observed}

\item{reciprocol}{whether to impute based on the reciprocal nature of the matrix}

\item{...}{passed to impute values probably need split2 = TRUE}
}
\description{
Impute Comix matrix based on Polymod
}
