% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_functions.R
\name{impute_values}
\alias{impute_values}
\title{Impute values of comix based on POLYMOD}
\usage{
impute_values(
  i = 0,
  home = TRUE,
  work = TRUE,
  other = TRUE,
  school = FALSE,
  ...
)
}
\arguments{
\item{i}{number of bootstraps, if zero then represent a single matrix}

\item{home}{logical default TRUE to include home for imputation}

\item{work}{logical default TRUE to include work for imputation}

\item{other}{logical default TRUE to include other for imputation}

\item{school}{ogical default FALSE to include school for imputation}
}
\description{
Tale comix contacts matrix and impute the values based on the scale of the
dominant eigenvalues between comix and polymod
}
