% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_functions.R
\name{max_eigen_ratio}
\alias{max_eigen_ratio}
\title{Eigenvalue ratio}
\usage{
max_eigen_ratio(x, y)
}
\arguments{
\item{x}{square matrix with the same dimentions as y}

\item{y}{square matrix with the same dimentions as y}
}
\description{
Take the ratio of the max eigenvalues of two matrices
}
