% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_questions.R
\name{reshape_questions}
\alias{reshape_questions}
\title{Reshapes the SPSS data format into multiple tables}
\usage{
reshape_questions(df, export_var_names = FALSE)
}
\arguments{
\item{df}{data.table object  Path to the input file}

\item{export_var_names}{Whether to export the variable names into a seperate file}
}
\value{
Multiple data.table object named as different tables each table refers to
a set of questions asked in the survey
}
\description{
This function take the survey data.table format that is from the spss file
and reshape and returns multiple objects with the prefix table
which relate to different questions asked in the survey
}
