% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_functions.R
\name{scale_factor}
\alias{scale_factor}
\title{Scale by eigenvalue ratio}
\usage{
scale_factor(mat1, mat2, split1 = FALSE, split2 = FALSE, ...)
}
\arguments{
\item{mat1}{Matrix}

\item{mat2}{Matrix}

\item{split1}{whether to split the first matrix}

\item{split2}{whether to split the second matrix}
}
\description{
Take two matrices split them if required to be of the same size and square
Then take the ratio of the maximum eigenvalues
}
