% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\name{sm_to_dt_matrix}
\alias{sm_to_dt_matrix}
\title{Converts socialmixR matrix into a dataframe for plotting}
\usage{
sm_to_dt_matrix(cmatrices)
}
\arguments{
\item{cmatrices}{list of socialmixR contact matrix objects}
}
\description{
This function takes a list of contact matrix objects from the socialmixR
package and returns a data.table object
}
