% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_functions.R
\name{split_cm}
\alias{split_cm}
\title{Splits a matrix by rows and columns}
\usage{
split_cm(mat, i = 1, row = 3:8, col = 3:8, ...)
}
\arguments{
\item{mat}{matrix or list of matrices}

\item{i}{the element in the list to use if mat is a list of matrices, defaults
to false for a single matrix}

\item{row}{a value or range (example, 1:3) of the rows to include}

\item{column}{a value or range (example, 1:3) of the rows to include}
}
\description{
Take a matrix or list of matrices and split then based on the row and column
values provided
}
