% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_functions.R
\name{split_cm_single}
\alias{split_cm_single}
\title{Will add documentation / may refactor}
\usage{
split_cm_single(mat, i = 1, row = 3:8, col = 3:8, ...)
}
\description{
Will add documentation / may refactor
}
