% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_functions.R
\name{update_cm}
\alias{update_cm}
\title{Will add documentation}
\usage{
update_cm(
  mat1,
  mat2,
  impute_rows = 1:2,
  impute_cols = 1:8,
  i = 1,
  observed_rows = 2:8,
  observed_col = 1:8,
  school = FALSE,
  ...
)
}
\description{
Will add documentation
}
