% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cm.R
\name{create_cm}
\alias{create_cm}
\title{Create a contacts matrix using socialmixr}
\usage{
create_cm(
  survey,
  matrix_output = TRUE,
  country = "United Kingdom",
  age_limits = c(0, 5, 18, 30, 40, 50, 60, 70),
  symmetric = FALSE,
  filter = FALSE,
  phys_contact = NULL,
  home = NULL,
  other_house = NULL,
  work = NULL,
  school = NULL,
  gender = NULL,
  hhm_quarantine = NULL,
  hhm_isolate = NULL,
  hhm_limit_work = NULL,
  hhm_limit_school = NULL,
  hhm_work_closed = NULL,
  part_quarantine = NULL,
  part_isolate = NULL,
  part_limit_work = NULL,
  part_limit_school = NULL,
  part_work_closed = NULL,
  part_covid_test_results = NULL,
  part_covid_contact = NULL,
  boots = 1,
  ...
)
}
\arguments{
\item{survey}{a survey object from socialmixr}

\item{matrix_output}{return just the matrix or the matrix and participants default = TRUE}

\item{country}{limit to one or more countries; if not given, will use all countries in the survey; these can be given as country names or 2-letter (ISO Alpha-2) country codes}

\item{age_limits}{lower limits of the age groups over which to construct the matrix}

\item{symmetric}{whether to make matrix symmetric}

\item{filter}{whether to filter the data default is FALSE, need to be TRUE for filter variables to work}

\item{phys_contact}{filter by physical or non-phyisical contact}

\item{home}{filter by contacts in the home}

\item{other_house}{filter to contacts in other homes}

\item{work}{filter by contacts at work}

\item{school}{fitler by contacts at school}

\item{gender}{filter by gender}

\item{hhm_quarantine}{filter by contacts with household members in quarantine}

\item{hhm_isolate}{filter by contacts with household members in isolation}

\item{hhm_limit_work}{filter by contacts with household members who have been asked to  limit work}

\item{hhm_limit_school}{filter by contacts with household members who have have been asked to limit school}

\item{hhm_work_closed}{filter by contacts with household members who have had their work closed}

\item{part_quarantine}{filter by participants in quarantine}

\item{part_isolate}{filter by participants in isolation}

\item{part_limit_work}{filter by participants who have been asked to limit work}

\item{part_limit_school}{filter by participants who have been asked to limit school}

\item{part_work_closed}{filter by participants who have had their work closed}

\item{part_covid_test_results}{filter by participants with a positive covid test result}

\item{part_covid_contact}{filter by participants who have had contact with a someone who has had covid}

\item{boots}{number of bootstrap sample}

\item{export_var_names}{Whether to export the variable names into a seperate file}
}
\value{
Either contact matrix or contact matrix and with participants information
}
\description{
Wrapper function for socialmix which helps quickly create filtered contacts matrices
and reshape and returns multiple objects with the prefix table
which relate to different questions asked in the survey
}
