% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_functions.R
\name{impute_cm}
\alias{impute_cm}
\title{Impute Comix matrix based on Polymod}
\usage{
impute_cm(
  comix,
  polymod,
  i = 0,
  impute_rows = 1:3,
  impute_cols = 1:3,
  observed_rows = 4:9,
  observed_cols = 1:9,
  reciprocol = TRUE,
  ...
)
}
\description{
Impute Comix matrix based on Polymod
}
