% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_functions.R
\name{scale_factor}
\alias{scale_factor}
\title{Scale by eigenvalue ratio}
\usage{
scale_factor(mat1, mat2, ...)
}
\arguments{
\item{mat1}{Matrix}

\item{mat2}{Matrix}

\item{...}{parameters to pass to the `split_cm` function}

\item{split1}{logical parameter, defaulted to FALSE, for splitting the first matrix}

\item{split2}{logical parameter, defaulted to FALSE, for splitting the second matrix}
}
\description{
Take two matrices split them if required to be of the same size and square
Then take the ratio of the maximum eigenvalues
}
