% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_functions.R
\name{scale_matrix}
\alias{scale_matrix}
\title{Scale by matrix by eigenvalue ratio}
\usage{
scale_matrix(fullmat, mat1, mat2, i = 1, ...)
}
\arguments{
\item{scale_matrix}{Full matrix which which is used for Matrix}

\item{missing_matrix}{Matrix with missing value that need imputing}

\item{split1}{whether to split the first matrix}

\item{split2}{whether to split the second matrix}
}
\description{
Take two matrices split them if required to be of the same size and square
Then take the ratio of the maximum eigenvalues
}
