% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_functions.R
\name{update_cm}
\alias{update_cm}
\title{Impute only the required values of the contact matrix}
\usage{
update_cm(
  mat1,
  mat2,
  impute_rows = 1:3,
  impute_cols = 1:9,
  i = 1,
  observed_rows = 4:9,
  observed_col = 1:9,
  reciprocol = TRUE,
  school = FALSE,
  ...
)
}
\description{
Only impute the required rows.
Will add documentation
}
